% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{dgm_var}
\alias{dgm_var}
\title{dgm_var}
\usage{
dgm_var(input, output, session, dgm)
}
\arguments{
\item{dgm}{a data.frame with names 'appear', 'disappear', and 'dim'.}
}
\value{
list with following components
\describe{
  \item{time_window}{reactive character string indicating x variable selection}
  \item{dims_selected}{reactive character string indicating y variable selection}
  \item{noise_threshold}{ }
}
}
\description{
dgm_var
}
