% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{linkedBarcode}
\alias{linkedBarcode}
\title{linkedBarcode}
\usage{
linkedBarcode(input, output, session, dgm, dgm_vars)
}
\arguments{
\item{dgm}{a (non-reactive) data.frame with names 'appear', 'disappear', and 'dim'.}

\item{dgm_vars}{list containing reactive expressions 'time_window', 'dims_selected', and 'noise_threshold'}
}
\value{
list with following components
\describe{
  \item{xvar}{reactive character string indicating x variable selection}
  \item{yvar}{reactive character string indicating y variable selection}
}
}
\description{
linkedBarcode
}
